New upstream version 20210510-93u+m-1.0.0~beta.1

+++ src/cmd/ksh93/tests/coprocess.sh
@@ -2,6 +2,7 @@
 #                                                                      #
 #               This software is part of the ast package               #
 #          Copyright (c) 1982-2012 AT&T Intellectual Property          #
+#          Copyright (c) 2020-2021 Contributors to ksh 93u+m           #
 #                      and is licensed under the                       #
 #                 Eclipse Public License, Version 1.0                  #
 #                    by AT&T Intellectual Property                     #
@@ -18,19 +19,8 @@
 #                                                                      #
 ########################################################################
 # test the behavior of co-processes
-function err_exit
-{
-	print -u2 -n "\t"
-	print -u2 -r ${Command}[$1]: "${@:2}"
-	let Errors+=1
-}
-alias err_exit='err_exit $LINENO'
 
-Command=${0##*/}
-integer Errors=0
-
-tmp=$(mktemp -dt) || { err_exit mktemp -dt failed; exit 1; }
-trap "cd /; rm -rf $tmp" EXIT
+. "${SHTESTS_COMMON:-${0%/*}/_common}"
 
 if	[[ -d /cygdrive ]]
 then	err_exit cygwin detected - coprocess tests disabled - enable at the risk of wedging your system
@@ -103,7 +93,7 @@ do
 	$cat |&
 	!
 	chmod +x $file
-	sleep 10 |&
+	sleep 1 |&
 	$file 2> /dev/null || err_exit "parent $cat coprocess prevents script coprocess"
 	exec 5<&p 6>&p
 	exec 5<&- 6>&-
@@ -113,15 +103,15 @@ do
 	cop=$!
 	exp=Done
 	print -p $'print hello | '$cat$'\nprint '$exp
-	read -t 5 -p
-	read -t 5 -p
+	read -t 1 -p
+	read -t 1 -p
 	got=$REPLY
 	if	[[ $got != $exp ]]
 	then	err_exit "${SHELL-ksh} $cat coprocess io failed -- got '$got', expected '$exp'"
 	fi
 	exec 5<&p 6>&p
 	exec 5<&- 6>&-
-	{ sleep 4; kill $cop; } 2>/dev/null &
+	{ sleep .4; kill $cop; } 2>/dev/null &
 	spy=$!
 	if	wait $cop 2>/dev/null
 	then	kill $spy 2>/dev/null
@@ -134,15 +124,15 @@ do
 	echo line2 | grep 'line1'
 	} |&
 	SECONDS=0 count=0
-	while	read -p -t 10 line
+	while	read -p -t 1 line
 	do	((count++))
 	done
-	if	(( SECONDS > 8 ))
+	if	(( SECONDS > .8 ))
 	then	err_exit "$cat coprocess read -p hanging (SECONDS=$SECONDS count=$count)"
 	fi
 	wait $!
 	
-	( sleep 3 |& sleep 1 && kill $!; sleep 1; sleep 3 |& sleep 1 && kill $! ) ||
+	( sleep 3 |& sleep .1 && kill $!; sleep .5; sleep 3 |& sleep .1 && kill $! ) ||
 		err_exit "$cat coprocess cleanup not working correctly"
 	{ : |& } 2>/dev/null ||
 		err_exit "subshell $cat coprocess lingers in parent"
@@ -160,7 +150,7 @@ do
 			wait $!
 		done
 		print
-	) 2>/dev/null | read -t 10 r
+	) 2>/dev/null | read -t 1 r
 	[[ $r == $e ]] || err_exit "$cat coprocess timing bug -- expected $e, got '$r'"
 	
 	r=
@@ -168,21 +158,21 @@ do
 		integer i
 		for ((i = 1; i <= N; i++))
 		do	print $i |&
-			sleep 0.01
+			sleep 0.001
 			r=$r$($cat <&p)
 			wait $!
 		done
 		print $r
-	) 2>/dev/null | read -t 10 r
+	) 2>/dev/null | read -t 1 r
 	[[ $r == $e ]] || err_exit "$cat coprocess command substitution bug -- expected $e, got '$r'"
 	
 	(
 		$cat |&
-		sleep 0.01
+		sleep 0.001
 		exec 6>&p
 		print -u6 ok
 		exec 6>&-
-		sleep 2
+		sleep .2
 		kill $! 2> /dev/null
 	) && err_exit "$cat coprocess with subshell would hang"
 	for sig in IOT ABRT
@@ -192,12 +182,12 @@ do
 					pid=\$!
 					trap "print TRAP" \$sig
 					(
-						sleep 2
+						sleep .2
 						kill -\$sig \$\$
-						sleep 2
+						sleep .2
 						kill -\$sig \$\$
 						kill \$pid
-						sleep 2
+						sleep .2
 						kill \$\$
 					) &
 					while	read -p || ((\$? > 256))
@@ -225,6 +215,11 @@ do
 	trap - TERM
 	[[ $sleep_pid ]] && kill $sleep_pid
 	
+# TODO: The /bin/cat iteration of this test is known to hang intermittently on Debian and
+# derived systems (such as Ubuntu). See: https://github.com/ksh93/ksh/issues/132
+# It is temporarily disabled here to avoid Github CI failures due to this known issue
+# while we work on other things. Export DEBUG_COPROCESS to include it in the tests.
+if [[ $cat != "$bincat" || -v DEBUG_COPROCESS ]]; then
 	trap 'sleep_pid=; kill $pid; err_exit "$cat coprocess 2 hung"' TERM
 	{ sleep 5; kill $$; } &
 	sleep_pid=$!
@@ -236,6 +231,7 @@ do
 	wait $pid 2> /dev/null
 	trap - TERM
 	[[ $sleep_pid ]] && kill $sleep_pid
+fi
 	
 	trap 'sleep_pid=; kill $pid; err_exit "$cat coprocess 3 hung"' TERM
 	{ sleep 5; kill $$; } &
